# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Buc", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/companies/{company_id}/areas/": {
      "get": {
        "tags": ["areas"],
        "summary": "Read Areas",
        "description": "Retrieve areas for a specific company.",
        "operationId": "areas-read_areas",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AreasPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["areas"],
        "summary": "Create Area",
        "description": "Create new area within a company.",
        "operationId": "areas-create_area",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AreaCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AreaPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/areas/{area_id}": {
      "get": {
        "tags": ["areas"],
        "summary": "Read Area By Id",
        "description": "Get a specific area by id.",
        "operationId": "areas-read_area_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          },
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Area Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AreaPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["areas"],
        "summary": "Update Area",
        "description": "Update an area.",
        "operationId": "areas-update_area",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          },
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Area Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AreaUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AreaPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["areas"],
        "summary": "Delete Area",
        "description": "Delete an area.",
        "operationId": "areas-delete_area",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          },
          {
            "name": "area_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Area Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AreaCreate": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "AreaCreate"
      },
      "AreaPublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "company_id": { "type": "string", "format": "uuid", "title": "Company Id" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" }
        },
        "type": "object",
        "required": ["name", "id", "company_id", "created_at", "updated_at"],
        "title": "AreaPublic"
      },
      "AreaUpdate": {
        "properties": {
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "AreaUpdate"
      },
      "AreasPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/AreaPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "AreasPublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" } }
      }
    }
  }
}